Parameterizes TFF simulation worker to allow no RREs on top of clients.
Desire to experiment with eager compilation to TensorFlow has been coming up in a few different contexts; opening up an extra parameter here should make such experimentation simple. I'm always loathe to add a Boolean parameter; the other clear option I see involves promoting the compositional pieces of executor stacks (which we have kept internal to this point) to a public position. I think executor stacks is still not in a state in which we want others taking dependencies on these pieces.

PiperOrigin-RevId: 348050964